6 commit

+++ calculator.py
@@ -1,5 +1,5 @@
-x = int(input("what is X?"))
-y = int(input("What is y?"))
+x = float(input("what is X?"))
+y = float(input("What is y?"))
 
 
 #z =int(x) + int(y)
